requirements.txt
python-dotenv~=1.1.0
Flask~=3.1.0
Flask-Login~=0.6.3
WTForms~=3.2.1
Flask-Bcrypt~=1.0.1
Flask-Migrate~=4.1.0
APScheduler~=3.10.1
Mako~=1.3.9
MarkupSafe~=3.0.2
pytz~=2025.2
typing_extensions~=4.13.0
Werkzeug~=3.1.3
click~=8.1.8
Jinja2~=3.1.6
blinker~=1.9.0
itsdangerous~=2.2.0
setuptools~=79.0.0
SQLAlchemy~=2.0.40
alembic~=1.15.1
tzlocal~=5.3.1
greenlet~=3.2.0
six~=1.17.0
bcrypt~=4.3.0
email-validator
python-telegram-bot~=20.3
dnspython~=2.7.0
idna~=3.10
httpcore~=0.17.3
httpx~=0.24.1
anyio~=4.9.0
h11~=0.14.0
wheel~=0.45.1
typing_extensions~=4.13.0
sniffio~=1.3.1
certifi~=2025.1.31
psycopg2-binary~=2.9.10